Wait for volumes creation/destruction after creating/removing the corresponding resources. Reorganize the Volumes code.
