+++ libvirt/resource_libvirt_volume_test.go
@@ -212,7 +212,7 @@ func TestAccLibvirtVolume_ManuallyDestroyed(t *testing.T) {
 					if err != nil {
 						panic(err)
 					}
-					removeVolume(client, id)
+					volumeDelete(client, id)
 				},
 			},
 		},
